Use pointers in mirrors map to allow easier update
The mirror monitor (in the following commit) need to update the mirror
status as it iterating the mirrors map.  The range() iterator always
returns a *copy* of the map key and value, so if one want to update the
elements, a map reassignment is required.  On ther other, use pointers
in the map looks better.