common: Fix log functions to report correct filename and line info
The log functions wrap around Go's builtin 'log' module.  While the
latter reports the filename and line where the log messages emit,
they were always the file 'log.go' and lines in it.

So implement the 'getOrigin()' function to obtain the real caller info
by our own.  In addition, obtain the function name of the caller and
report in the log messages.

Credit: https://stackoverflow.com/a/58232098